+++ lib/src/builders/guild/scheduled_event.dart
@@ -24,22 +24,59 @@ class ScheduledEventBuilder extends CreateBuilder<ScheduledEvent> {
 
   ImageBuilder? image;
 
+  RecurrenceRuleBuilder? recurrenceRule;
+
   ScheduledEventBuilder({
-    required this.channelId,
+    this.channelId,
     this.metadata,
     required this.name,
     required this.privacyLevel,
     required this.scheduledStartTime,
-    required this.scheduledEndTime,
+    this.scheduledEndTime,
     this.description,
     required this.type,
     this.image,
+    this.recurrenceRule,
   });
+
+  ScheduledEventBuilder.stageInstance({
+    required Snowflake this.channelId,
+    required this.name,
+    required this.privacyLevel,
+    required this.scheduledStartTime,
+    this.scheduledEndTime,
+    this.description,
+    this.image,
+    this.recurrenceRule,
+  }) : type = ScheduledEntityType.stageInstance;
+
+  ScheduledEventBuilder.voice({
+    required Snowflake this.channelId,
+    required this.name,
+    required this.privacyLevel,
+    required this.scheduledStartTime,
+    this.scheduledEndTime,
+    this.description,
+    this.image,
+    this.recurrenceRule,
+  }) : type = ScheduledEntityType.voice;
+
+  ScheduledEventBuilder.external({
+    required this.name,
+    required this.privacyLevel,
+    required this.scheduledStartTime,
+    required DateTime this.scheduledEndTime,
+    required String location,
+    this.description,
+    this.image,
+    this.recurrenceRule,
+  })  : type = ScheduledEntityType.external,
+        metadata = EntityMetadata(location: location);
 
   @override
   Map<String, Object?> build() => {
         if (channelId != null) 'channel_id': channelId.toString(),
-        if (metadata != null) 'metadata': {'location': metadata!.location},
+        if (metadata != null) 'entity_metadata': {'location': metadata!.location},
         'name': name,
         'privacy_level': privacyLevel.value,
         'scheduled_start_time': scheduledStartTime.toIso8601String(),
@@ -47,6 +84,7 @@ class ScheduledEventBuilder extends CreateBuilder<ScheduledEvent> {
         if (description != null) 'description': description,
         'entity_type': type.value,
         if (image != null) 'image': image!.buildDataString(),
+        if (recurrenceRule != null) 'recurrence_rule': recurrenceRule!.build(),
       };
 }
 
@@ -71,6 +109,8 @@ class ScheduledEventUpdateBuilder extends UpdateBuilder<ScheduledEvent> {
 
   ImageBuilder? image;
 
+  RecurrenceRuleBuilder? recurrenceRule;
+
   ScheduledEventUpdateBuilder({
     this.channelId = sentinelSnowflake,
     this.metadata = sentinelEntityMetadata,
@@ -82,6 +122,7 @@ class ScheduledEventUpdateBuilder extends UpdateBuilder<ScheduledEvent> {
     this.type,
     this.status,
     this.image,
+    this.recurrenceRule,
   });
 
   @override
@@ -96,5 +137,61 @@ class ScheduledEventUpdateBuilder extends UpdateBuilder<ScheduledEvent> {
         if (type != null) 'entity_type': type!.value,
         if (status != null) 'status': status!.value,
         if (image != null) 'image': image!.buildDataString(),
+        if (recurrenceRule != null) 'recurrence_rule': recurrenceRule!.build(),
+      };
+}
+
+class RecurrenceRuleBuilder extends CreateBuilder<RecurrenceRule> {
+  DateTime start;
+  RecurrenceRuleFrequency frequency;
+  int interval;
+  List<RecurrenceRuleWeekday>? byWeekday;
+  List<RecurrenceRuleNWeekday>? byNWeekday;
+  List<RecurrenceRuleMonth>? byMonth;
+  List<int>? byMonthDay;
+
+  RecurrenceRuleBuilder({
+    required this.start,
+    required this.frequency,
+    required this.interval,
+    this.byWeekday,
+    this.byNWeekday,
+    this.byMonth,
+    this.byMonthDay,
+  });
+
+  RecurrenceRuleBuilder.daily({required this.start, this.byWeekday})
+      : frequency = RecurrenceRuleFrequency.daily,
+        interval = 1;
+
+  RecurrenceRuleBuilder.weekly({
+    required this.start,
+    required this.interval,
+    RecurrenceRuleWeekday? day,
+  })  : frequency = RecurrenceRuleFrequency.weekly,
+        byWeekday = day == null ? null : [day];
+
+  RecurrenceRuleBuilder.monthly({
+    required this.start,
+    RecurrenceRuleNWeekday? day,
+  })  : frequency = RecurrenceRuleFrequency.monthly,
+        interval = 1,
+        byNWeekday = day == null ? null : [day];
+
+  RecurrenceRuleBuilder.yearly({required this.start, (RecurrenceRuleMonth, int)? day})
+      : frequency = RecurrenceRuleFrequency.yearly,
+        interval = 1,
+        byMonth = day == null ? null : [day.$1],
+        byMonthDay = day == null ? null : [day.$2];
+
+  @override
+  Map<String, Object?> build() => {
+        'start': start.toIso8601String(),
+        'frequency': frequency.value,
+        'interval': interval,
+        'by_weekday': byWeekday?.map((weekday) => weekday.value).toList(),
+        'by_n_weekday': byNWeekday?.map((nWeekday) => {'n': nWeekday.n, 'day': nWeekday.day.value}).toList(),
+        'by_month': byMonth?.map((month) => month.value).toList(),
+        'by_month_day': byMonthDay,
       };
 }