+++ lib/src/models/application.dart
@@ -1,7 +1,10 @@
 import 'package:nyxx/src/http/cdn/cdn_asset.dart';
 import 'package:nyxx/src/http/managers/application_manager.dart';
+import 'package:nyxx/src/http/managers/emoji_manager.dart';
 import 'package:nyxx/src/http/managers/entitlement_manager.dart';
+import 'package:nyxx/src/http/managers/sku_manager.dart';
 import 'package:nyxx/src/http/route.dart';
+import 'package:nyxx/src/models/emoji.dart';
 import 'package:nyxx/src/models/guild/guild.dart';
 import 'package:nyxx/src/models/locale.dart';
 import 'package:nyxx/src/models/permissions.dart';
@@ -25,6 +28,12 @@ class PartialApplication with ToStringHelper {
 
   /// An [EntitlementManager] for this application's [Entitlement]s.
   EntitlementManager get entitlements => EntitlementManager(manager.client.options.entitlementConfig, manager.client, applicationId: id);
+
+  /// An [ApplicationEmojiManager] for this application's [Emoji]s.
+  ApplicationEmojiManager get emojis => ApplicationEmojiManager(manager.client.options.emojiCacheConfig, manager.client, applicationId: id);
+
+  /// An [SkuManager] for this application's [Sku]s.
+  SkuManager get skus => SkuManager(manager.client.options.skuConfig, manager.client, applicationId: id);
 
   /// Create a new [PartialApplication].
   /// @nodoc
@@ -132,6 +141,9 @@ class Application extends PartialApplication {
   /// When configured, this will render the app as a verification method in the guild role verification configuration.
   final Uri? roleConnectionsVerificationUrl;
 
+  /// The approximate number of users that have installed this application.
+  final int? approximateUserInstallCount;
+
   /// {@macro application}
   /// @nodoc
   Application({
@@ -163,6 +175,7 @@ class Application extends PartialApplication {
     required this.integrationTypesConfig,
     required this.customInstallUrl,
     required this.roleConnectionsVerificationUrl,
+    required this.approximateUserInstallCount,
   });
 
   /// This application's icon.